Estimate alpha diversity

Estimates a number of standard alpha diversity in a statistically rigourous way, and returns the results as a data.frame. Strictly speaking, this function is not only estimating richness, despite its name. It can operate on the cumulative population of all samples in the dataset, or by repeating the richness estimates for each sample individually. NOTE: You must use untrimmed datasets for meaningful results, as these estimates (and even the ``observed'' richness) are highly dependent on the number of singletons. You can always trim the data later on if needed, just not before using this function.

estimate_alpha_better(physeq, split = TRUE, measures = NULL,
  estimators = NULL)

Arguments

physeq
(Required). phyloseq-class, or alternatively, an otu_table-class. The data about which you want to estimate the richness.

split
(Optional). Logical. Should a separate set of richness estimates be performed for each sample? Or alternatively, pool all samples and estimate richness of the entire set.

measures
(Optional). Default is NULL, meaning that all available alpha-diversity measures will be included. Alternatively, you can specify one or more measures as a character vector of measure names. Values must be among those supported: c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher").

Value

A data.frame of the richness estimates, and their standard error.

See also