Species richness estimation with Kemp-type models

This function implements the species richness estimation procedure outlined in Willis & Bunge (2015). The diversity estimate, standard error, estimated model coefficients, model details and plot of the fitted model are returned.

kemp(
  input_data,
  cutoff = NA,
  output = NULL,
  plot = NULL,
  answers = NULL,
  print = NULL,
  ...
)

Arguments

input_data
An input type that can be processed by convert()

cutoff
The maximum frequency count to use for model fitting

output
Deprecated; only for backwards compatibility

plot
Deprecated; only for backwards compatibility

answers
Deprecated; only for backwards compatibility

print
Deprecated; only for backwards compatibility

...
Additional arguments will be ignored; this is for backward compatibility

Value

An object of class alpha_estimate

code

A category representing algorithm behaviour. code=1 indicates no nonlinear models converged and the transformed WLRM diversity estimate of Rocchetti et. al. (2011) is returned. code=2 indicates that the iteratively reweighted model converged and was returned. code=3 indicates that iterative reweighting did not converge but a model based on a simplified variance structure was returned (in this case, the variance of the frequency ratios is assumed to be proportional to the denominator frequency index). Please peruse your fitted model before using your diversity estimate.

name

The ``name'' of the selected model. The first integer represents the numerator polynomial degree and the second integer represents the denominator polynomial degree of the model for the frequency ratios. See Willis & Bunge (2015) for details.

para

Estimated model parameters and standard errors.

est

The estimate of total (observed plus unobserved) diversity.

seest

The standard error in the diversity estimate.

full

The chosen nonlinear model for frequency ratios.

ci

An asymmetric 95% confidence interval for diversity.

References

Willis, A. and Bunge, J. (2015). Estimating diversity via frequency ratios. Biometrics, 71(4), 1042--1049.

Rocchetti, I., Bunge, J. and Bohning, D. (2011). Population size estimation based upon ratios of recapture probabilities. Annals of Applied Statistics, 5.

See also

Author

Amy Willis

Examples

kemp(apples)
#> Estimate of richness from method kemp:
#>   Estimate is 1552
#>  with standard error 304.71
#>   Confidence interval: (1007, 47805)
#> 
kemp(apples, plot = FALSE, output = FALSE, answers = TRUE)
#> Estimate of richness from method kemp:
#>   Estimate is 1552
#>  with standard error 304.71
#>   Confidence interval: (1007, 47805)
#>