Objective Bayes species richness estimate with the mixed-geometric model

Objective Bayes species richness estimate with the mixed-geometric model

objective_bayes_mixedgeo(
  data,
  output = TRUE,
  plot = TRUE,
  answers = FALSE,
  tau = 10,
  burn.in = 100,
  iterations = 2500,
  Metropolis.stdev.N = 100,
  Metropolis.start.T1 = 1,
  Metropolis.stdev.T1 = 2,
  Metropolis.start.T2 = 3,
  Metropolis.stdev.T2 = 2,
  bars = 3
)

Arguments

data

A data.frame or tibble containing the frequency of frequencies data.

output

A logical value indicating whether the function should return the full posterior samples and diagnostic plots. Defaults to TRUE.

plot

A logical value indicating whether to print a plot of the Markov Chain Monte Carlo diagnostic chains and posterior distribution. Defaults to TRUE.

answers

A logical value indicating whether to return print-ready summary statistics and credible intervals. Defaults to FALSE.

tau

An integer indicating the maximum frequency of the data to be analyzed. Defaults to 10.

burn.in

An integer indicating the number of burn-in iterations for the Markov Chain Monte Carlo sampler. Defaults to 100.

iterations

An integer indicating the number of additional iterations for the Markov Chain Monte Carlo sampler after the burn-in iterations. Defaults to 2500.

Metropolis.stdev.N

A numeric value specifying the standard deviation for the Metropolis-Hastings proposal distribution of the parameter N. Defaults to 100.

Metropolis.start.T1

Starting value in the Metropolis algorithm for the parameter theta1. Defaults to 1.

Metropolis.stdev.T1

A numeric value specifying the standard deviation for the Metropolis-Hastings proposal distribution of the parameter theta1. Defaults to 2.

Metropolis.start.T2

Starting value in the Metropolis algorithm for the parameter theta2. Defaults to 3.

Metropolis.stdev.T2

A numeric value specifying the standard deviation for the Metropolis-Hastings proposal distribution of the parameter theta2. Defaults to 2.

bars

An integer indicating the width of bars in the histogram contained in output. Defaults to 3.

Value

A list of results, including

est

the median of estimates of N

,

ci

a confidence interval for N

,

mean

the mean of estimates of N

,

semeanest

the standard error of mean estimates

,

dic

the DIC of the model

,

fits

fitted values

, and

diagnostics

model diagonstics

.