Plot function for alpha_estimates class

Plot function for alpha_estimates class

# S3 method for alpha_estimates
plot(
  x,
  physeq = NULL,
  measure = NULL,
  color = NULL,
  shape = NULL,
  title = NULL,
  trim_plot = FALSE,
  ...
)

Arguments

x
Object of class alpha_estimates.

physeq
(Optional). Default NULL. Required object of class phyloseq if including a sample_data variable for color or shape.

measure
(Optional). If there are multiple richness measures included in x, this can be set to the the desired measure to be plotted. Defaults to the measure of the first alpha diversity estimate.

color
(Optional). Default NULL. The sample variable to map to different colors. Can be a single character string of the variable name in sample_data or a custom supplied vector with length equal to the number of samples.

shape
(Optional). Default NULL. The sample variable to map to different shapes. Can be a single character string of the variable name in sample_data or a custom supplied vector with length equal to the number of samples.

title
(Optional). Default NULL. Character string. The main title for the graphic.

trim_plot
(Optional). Default FALSE. Boolean indicator for whether you want the plot to include the full confidence intervals.

...
See details

Details

... does not currently have any implemented options. Optional arguments currently include "trim_plot", a Optional

Examples

if (FALSE) {
library(phyloseq)
data(GlobalPatterns)
alphas <- breakaway(GlobalPatterns)
plot(alphas)
}