Plot alpha diversity with less incorrect error bars

For no good reason, we almost never plot alpha diversity with its appropriate standard errors. This function is intended to fix that.

plot_alpha(physeq, x = "samples", color = NULL, shape = NULL, nrow = 1,
  measures = NULL, estimators = NULL, upper_limits = NULL)

Arguments

physeq
(Required). phyloseq-class, or alternatively, an otu_table-class. The data about which you want to estimate.

x
(Optional). A variable to map to the horizontal axis. The vertical axis will be mapped to the alpha diversity index/estimate and have units of total taxa, and/or index value (dimensionless). This parameter (x) can be either a character string indicating a variable in sample_data (among the set returned by sample_variables(physeq) ); or a custom supplied vector with length equal to the number of samples in the dataset (nsamples(physeq)).

The default value is "samples", which will map each sample's name to a separate horizontal position in the plot.

color
(Optional). Default NULL. The sample variable to map to different colors. Like x, this can be a single character string of the variable name in sample_data (among the set returned by sample_variables(physeq) ); or a custom supplied vector with length equal to the number of samples in the dataset (nsamples(physeq)). The color scheme is chosen automatically by link{ggplot}, but it can be modified afterward with an additional layer using scale_color_manual.

shape
(Optional). Default NULL. The sample variable to map to different shapes. Like x and color, this can be a single character string of the variable name in sample_data (among the set returned by sample_variables(physeq) ); or a custom supplied vector with length equal to the number of samples in the dataset (nsamples(physeq)). The shape scale is chosen automatically by link{ggplot}, but it can be modified afterward with an additional layer using scale_shape_manual.

nrow
(Optional). Default is 1, meaning that all plot panels will be placed in a single row, side-by-side. This argument is passed to facet_wrap. If NULL, the number of rows and columns will be chosen automatically (wrapped) based on the number of panels and the size of the graphics device.

measures
(Optional). Default is NULL, meaning that all available alpha-diversity measures will be included in plot panels. Alternatively, you can specify one or more measures as a character vector of measure names. Values must be among those supported: c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher").

title
(Optional). Default NULL. Character string. The main title for the graphic.

Value

A ggplot plot object summarizing the richness estimates, and their standard error.

Details

NOTE: This code is modify from the phyloseq functions plot_richness and estimate_richness. Please cite phyloseq as well as breakaway if you use this function!

See also