Species richness estimation with breakaway

This function implements the species richness estimation procedure outlined in Willis & Bunge (2015). The diversity estimate, standard error, estimated model coefficients, model details and plot of the fitted model are returned.

breakaway(my_data, print = TRUE, plot = FALSE, answers = FALSE, force = FALSE,
    useAll = FALSE)

Arguments

my_data
The sample frequency count table for the population of interest. The first row must correspond to the singletons. Acceptable formats include a matrix, data frame, or file path (csv or txt). The standard frequency count table format is used: two columns, the first of which contains the frequency of interest (eg. 1 for singletons, species observed once, 2 for doubletons, species observed twice, etc.) and the second of which contains the number of species observed this many times. Frequencies (first column) should be ordered least to greatest. At least 6 contiguous frequencies are necessary. Do not concatenate large frequencies. See dataset apples for sample formatting.

print
Logical: whether the results should be printed to screen. If FALSE, answers should be set to TRUE so that results will be returned.

plot
Logical: whether the data and model fit should be plotted.

answers
Logical: whether the function should return an argument. If FALSE, print should be set to TRUE.

force
Logical: force breakaway to run in the presence of frequency count concatenation, i.e. combining all species observed 27 or 28 or 29 or more into frequency count 27. force=TRUE will force breakaway to fit models in the presence of this. This will generally result in the model being misspecified, and breakaway's diversity estimates cannot be considered reliable in this case. The option force issue does not address censoring or small data on the rare end of the frequency count tables, i.e. insufficiently many contiguous frequencies.

useAll
Forces all data to be used. Otherwise breakaway determines the optimal cutoff.

Value

code

A category representing algorithm behaviour. code=1 indicates no nonlinear models converged and the transformed WLRM diversity estimate of Rocchetti et. al. (2011) is returned. code=2 indicates that the iteratively reweighted model converged and was returned. code=3 indicates that iterative reweighting did not converge but a model based on a simplified variance structure was returned (in this case, the variance of the frequency ratios is assumed to be proportional to the denominator frequency index). Please peruse your fitted model before using your diversity estimate.

name

The ``name'' of the selected model. The first integer represents the numerator polynomial degree and the second integer represents the denominator polynomial degree of the model for the frequency ratios. See Willis & Bunge (2015) for details.

para

Estimated model parameters and standard errors.

est

The estimate of total (observed plus unobserved) diversity.

seest

The standard error in the diversity estimate.

full

The chosen nonlinear model for frequency ratios.

ci

An asymmetric 95% confidence interval for diversity.

Note

breakaway presents an estimator of species richness that is well-suited to the high-diversity/microbial setting. However, many microbial datasets display more diversity than the Kemp-type models can permit. In this case, the log-transformed WLRM diversity estimator of Rocchetti et. al. (2011) is returned. The authors' experience suggests that some datasets that require the log-transformed WLRM contain ``false'' diversity, that is, diversity attributable to sequencing errors (via an inflated singleton count). The authors encourage judicious use of diversity estimators when the dataset may contain these errors, and recommend the use of breakaway_nof1 as an exploratory tool in this case.

References

Willis, A. and Bunge, J. (2015). Estimating diversity via frequency ratios. Biometrics, 71(4), 1042--1049.

Rocchetti, I., Bunge, J. and Bohning, D. (2011). Population size estimation based upon ratios of recapture probabilities. Annals of Applied Statistics, 5.

See also

breakaway_nof1; apples

Examples

breakaway(apples)

#> We used 1/x weighting. 
#> ################## breakaway ##################
#> 	The best estimate of total diversity is 1552 
#>  	 with std error 305 
#> 	The model employed was model_1_1 
#> 	The function selected was
#> 	  f_{x+1}/f_{x} ~ (beta0+beta1*(x-xbar))/(1+alpha1*(x-xbar)) 
#>        Coef estimates Coef std errors
#> beta0      1.20345572      0.16807523
#> beta1      0.05765149      0.02962840
#> alpha1     0.03012304      0.03782164
#> xbar			 16.5

breakaway(apples, plot = FALSE, print = FALSE, answers = TRUE)

#> $code
#> [1] 3
#> 
#> $name
#> [1] "model_1_1"
#> 
#> $para
#>        Coef estimates Coef std errors
#> beta0      1.20345572      0.16807523
#> beta1      0.05765149      0.02962840
#> alpha1     0.03012304      0.03782164
#> 
#> $est
#> [1] 1552.416
#> 
#> $seest
#> [1] 304.7071
#> 
#> $full
#> Nonlinear regression model
#>   model: lhs$y ~ structure_1_1(x, beta0, beta1, alpha1)
#>    data: lhs
#>   beta0   beta1  alpha1 
#> 1.20346 0.05765 0.03012 
#>  weighted residual sum-of-squares: 1.274
#> 
#> Number of iterations to convergence: 8 
#> Achieved convergence tolerance: 8.599e-06
#> 
#> $ci
#> [1]  1006.52 47805.11
#>