species richness estimation without singletons

This function permits estimation of total diversity based on a sample frequency count table. Unlike breakaway, it does not require an input for the number of species observed once, making it an excellent exploratory tool for microbial ecologists who believe that their sample may contain spurious singletons. The underlying estimation procedure is similar to that of breakaway and is outlined in Willis & Bunge (2014). The diversity estimate, standard error, estimated model coefficients and plot of the fitted model are returned.

breakaway_nof1(
  input_data,
  output = NULL,
  plot = NULL,
  answers = NULL,
  print = NULL
)

Arguments

input_data
An input type that can be processed by convert()

output
Deprecated; only for backwards compatibility

plot
Deprecated; only for backwards compatibility

answers
Deprecated; only for backwards compatibility

print
Deprecated; only for backwards compatibility

Value

An object of class alpha_estimate

code

A category representing algorithm behaviour. code=1 indicates no nonlinear models converged and the transformed WLRM diversity estimate of Rocchetti et. al. (2011) is returned. code=2 indicates that the iteratively reweighted model converged and was returned. code=3 indicates that iterative reweighting did not converge but a model based on a simplified variance structure was returned (in this case, the variance of the frequency ratios is assumed to be proportional to the denominator frequency index). Please peruse your fitted model before using your diversity estimate.

name

The ``name'' of the selected model. The first integer represents the numerator polynomial degree and the second integer represents the denominator polynomial degree. See Willis & Bunge (2014) for details.

para

Estimated model parameters and standard errors.

est

The estimate of total (observed plus unobserved) diversity.

seest

The standard error in the diversity estimate.

full

The chosen nonlinear model for frequency ratios.

Note

It is common for microbial ecologists to believe that their dataset contains false diversity. This often arises because sequencing errors result in classifying already observed organisms as new organisms. breakaway_nof1 was developed as an exploratory tool in this case. Practitioners can run breakaway on their dataset including the singletons, and breakaway_nof1 on their dataset excluding the singletons, and assess if the estimated levels of diversity are very different. Great disparity may provide evidence of an inflated singleton count, or at the very least, that breakaway is especially sensitive to the number of rare species observed. Note that breakaway_nof1 may be less stable than breakaway due to predicting based on a reduced dataset, and have greater standard errors.

References

Willis, A. (2015). Species richness estimation with high diversity but spurious singletons. arXiv.

Willis, A. and Bunge, J. (2015). Estimating diversity via frequency ratios. Biometrics.

See also

Author

Amy Willis

Examples

breakaway_nof1(apples[-1, ])

#> Estimate of richness from method breakaway_nof1:
#>   Estimate is 1500
#>  with standard error 1341.35
#>   Confidence interval: (725, 139033)
#> 

breakaway_nof1(apples[-1, ], plot = FALSE, output = FALSE, answers = TRUE)

#> Estimate of richness from method breakaway_nof1:
#>   Estimate is 1500
#>  with standard error 1341.35
#>   Confidence interval: (725, 139033)
#>