Chao-Bunge species richness estimator

This function implements the species richness estimation procedure outlined in Chao & Bunge (2002).

chao_bunge(data, cutoff = 10, print = TRUE, answers = FALSE)

Arguments

data
The sample frequency count table for the population of interest. See dataset apples for sample formatting.

cutoff
The maximum frequency to use in fitting.

print
Logical: whether the results should be printed to screen.

answers
Should the answers be returned as a list?

Value

The results of the estimator, including standard error.

Examples

chao_bunge(apples)

#>       [,1]     [,2]      [,3]      [,4]      [,5]      [,6]      [,7]      [,8]
#>  [1,]    0 -31.9861 -25.74982 -11.86906 -8.449159 -10.25969 -5.230431 -5.632772
#>  [2,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [3,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [4,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [5,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [6,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [7,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [8,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>  [9,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#> [10,]    0   0.0000   0.00000   0.00000  0.000000   0.00000  0.000000  0.000000
#>            [,9]    [,10]
#>  [1,] -2.615216 -4.62692
#>  [2,]  0.000000  0.00000
#>  [3,]  0.000000  0.00000
#>  [4,]  0.000000  0.00000
#>  [5,]  0.000000  0.00000
#>  [6,]  0.000000  0.00000
#>  [7,]  0.000000  0.00000
#>  [8,]  0.000000  0.00000
#>  [9,]  0.000000  0.00000
#> [10,]  0.000000  0.00000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]      [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.230431
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.002305
#>  [3,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [4,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>            [,8]      [,9]     [,10]
#>  [1,] -5.632772 -2.615216 -4.626920
#>  [2,] -3.233252 -1.501153 -2.655886
#>  [3,]  0.000000  0.000000  0.000000
#>  [4,]  0.000000  0.000000  0.000000
#>  [5,]  0.000000  0.000000  0.000000
#>  [6,]  0.000000  0.000000  0.000000
#>  [7,]  0.000000  0.000000  0.000000
#>  [8,]  0.000000  0.000000  0.000000
#>  [9,]  0.000000  0.000000  0.000000
#> [10,]  0.000000  0.000000  0.000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]      [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.230431
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.002305
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.416950
#>  [4,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>            [,8]      [,9]     [,10]
#>  [1,] -5.632772 -2.615216 -4.626920
#>  [2,] -3.233252 -1.501153 -2.655886
#>  [3,] -2.602870 -1.208475 -2.138071
#>  [4,]  0.000000  0.000000  0.000000
#>  [5,]  0.000000  0.000000  0.000000
#>  [6,]  0.000000  0.000000  0.000000
#>  [7,]  0.000000  0.000000  0.000000
#>  [8,]  0.000000  0.000000  0.000000
#>  [9,]  0.000000  0.000000  0.000000
#> [10,]  0.000000  0.000000  0.000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]      [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.230431
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.002305
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.416950
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.114063
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.000000
#>            [,8]       [,9]      [,10]
#>  [1,] -5.632772 -2.6152157 -4.6269202
#>  [2,] -3.233252 -1.5011527 -2.6558856
#>  [3,] -2.602870 -1.2084751 -2.1380714
#>  [4,] -1.199760 -0.5570315 -0.9855173
#>  [5,]  0.000000  0.0000000  0.0000000
#>  [6,]  0.000000  0.0000000  0.0000000
#>  [7,]  0.000000  0.0000000  0.0000000
#>  [8,]  0.000000  0.0000000  0.0000000
#>  [9,]  0.000000  0.0000000  0.0000000
#> [10,]  0.000000  0.0000000  0.0000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]       [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.2304315
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.0023054
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.4169503
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.1140630
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000  -1.555621 -0.7930618
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>             [,8]       [,9]      [,10]
#>  [1,] -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.8540666 -0.3965309 -0.7015547
#>  [6,]  0.0000000  0.0000000  0.0000000
#>  [7,]  0.0000000  0.0000000  0.0000000
#>  [8,]  0.0000000  0.0000000  0.0000000
#>  [9,]  0.0000000  0.0000000  0.0000000
#> [10,]  0.0000000  0.0000000  0.0000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]       [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.2304315
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.0023054
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.4169503
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.1140630
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000  -1.555621 -0.7930618
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000 -0.9630036
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>             [,8]       [,9]      [,10]
#>  [1,] -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.8540666 -0.3965309 -0.7015547
#>  [6,] -1.0370808 -0.4815018 -0.8518878
#>  [7,]  0.0000000  0.0000000  0.0000000
#>  [8,]  0.0000000  0.0000000  0.0000000
#>  [9,]  0.0000000  0.0000000  0.0000000
#> [10,]  0.0000000  0.0000000  0.0000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]       [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.2304315
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.0023054
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.4169503
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.1140630
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000  -1.555621 -0.7930618
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000 -0.9630036
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>             [,8]       [,9]      [,10]
#>  [1,] -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.8540666 -0.3965309 -0.7015547
#>  [6,] -1.0370808 -0.4815018 -0.8518878
#>  [7,] -0.5287079 -0.2454715 -0.4342958
#>  [8,]  0.0000000  0.0000000  0.0000000
#>  [9,]  0.0000000  0.0000000  0.0000000
#> [10,]  0.0000000  0.0000000  0.0000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]       [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.2304315
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.0023054
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.4169503
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.1140630
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000  -1.555621 -0.7930618
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000 -0.9630036
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>             [,8]       [,9]      [,10]
#>  [1,] -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.8540666 -0.3965309 -0.7015547
#>  [6,] -1.0370808 -0.4815018 -0.8518878
#>  [7,] -0.5287079 -0.2454715 -0.4342958
#>  [8,]  0.0000000 -0.2643539 -0.4677031
#>  [9,]  0.0000000  0.0000000  0.0000000
#> [10,]  0.0000000  0.0000000  0.0000000
#>       [,1]     [,2]      [,3]       [,4]      [,5]       [,6]       [,7]
#>  [1,]    0 -31.9861 -25.74982 -11.869056 -8.449159 -10.259693 -5.2304315
#>  [2,]    0   0.0000 -14.78058  -6.812924 -4.849878  -5.889138 -3.0023054
#>  [3,]    0   0.0000   0.00000  -5.484618 -3.904304  -4.740941 -2.4169503
#>  [4,]    0   0.0000   0.00000   0.000000 -1.799640  -2.185277 -1.1140630
#>  [5,]    0   0.0000   0.00000   0.000000  0.000000  -1.555621 -0.7930618
#>  [6,]    0   0.0000   0.00000   0.000000  0.000000   0.000000 -0.9630036
#>  [7,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [8,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>  [9,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#> [10,]    0   0.0000   0.00000   0.000000  0.000000   0.000000  0.0000000
#>             [,8]       [,9]      [,10]
#>  [1,] -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.8540666 -0.3965309 -0.7015547
#>  [6,] -1.0370808 -0.4815018 -0.8518878
#>  [7,] -0.5287079 -0.2454715 -0.4342958
#>  [8,]  0.0000000 -0.2643539 -0.4677031
#>  [9,]  0.0000000  0.0000000 -0.2171479
#> [10,]  0.0000000  0.0000000  0.0000000
#>             [,1]       [,2]       [,3]        [,4]       [,5]        [,6]
#>  [1,] 221.275788 -31.986100 -25.749817 -11.8690560 -8.4491585 -10.2596925
#>  [2,] -31.986100 140.639748 -14.780581  -6.8129239 -4.8498780  -5.8891376
#>  [3,] -25.749817 -14.780581 116.101168  -5.4846180 -3.9043043  -4.7409409
#>  [4,] -11.869056  -6.812924  -5.484618  56.4719339 -1.7996403  -2.1852775
#>  [5,]  -8.449159  -4.849878  -3.904304  -1.7996403 40.7189001  -1.5556212
#>  [6,] -10.259693  -5.889138  -4.740941  -2.1852775 -1.5556212  49.1110313
#>  [7,]  -5.230431  -3.002305  -2.416950  -1.1140630 -0.7930618  -0.9630036
#>  [8,]  -5.632772  -3.233252  -2.602870  -1.1997602 -0.8540666  -1.0370808
#>  [9,]  -2.615216  -1.501153  -1.208475  -0.5570315 -0.3965309  -0.4815018
#> [10,]  -4.626920  -2.655886  -2.138071  -0.9855173 -0.7015547  -0.8518878
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] -5.2304315 -5.6327724 -2.6152157 -4.6269202
#>  [2,] -3.0023054 -3.2332520 -1.5011527 -2.6558856
#>  [3,] -2.4169503 -2.6028695 -1.2084751 -2.1380714
#>  [4,] -1.1140630 -1.1997602 -0.5570315 -0.9855173
#>  [5,] -0.7930618 -0.8540666 -0.3965309 -0.7015547
#>  [6,] -0.9630036 -1.0370808 -0.4815018 -0.8518878
#>  [7,] 25.5090570 -0.5287079 -0.2454715 -0.4342958
#>  [8,] -0.5287079 27.4306223 -0.2643539 -0.4677031
#>  [9,] -0.2454715 -0.2643539 12.8772642 -0.2171479
#> [10,] -0.4342958 -0.4677031 -0.2171479 22.6158153
#> ################## Chao-Bunge ##################
#> 	The estimate of total diversity is 1377 
#>  	 with std error 64