function for species richness estimation

This function implements the transformed version of the species richness estimation procedure outlined in Rocchetti, Bunge and Bohning (2011).

wlrm_transformed(data, print = TRUE, plot = FALSE, answers = FALSE, cutoff = NA)

Arguments

data
The sample frequency count table for the population of interest. See breakaway for details.

print
Logical: whether the results should be printed to screen.

plot
Logical: whether the data and model fit should be plotted.

answers
Logical: whether the function should return an argument.

cutoff
The maximum frequency count to use for prediction. Defaults to maximal.

Value

name

The ``name'' of the model.

para

Estimated model parameters and standard errors.

est

The estimate of total (observed plus unobserved) diversity.

seest

The standard error in the diversity estimate.

full

The chosen linear model for frequency ratios.

ci

An asymmetric 95% confidence interval for diversity.

Note

While robust to many different structures, model is almost always misspecified. The result is usually implausible diversity estimates with artificially low standard errors. Extreme caution is advised.

References

Rocchetti, I., Bunge, J. and Bohning, D. (2011). Population size estimation based upon ratios of recapture probabilities. Annals of Applied Statistics, 5.

See also

breakaway; apples; wlrm_untransformed

Examples

wlrm_transformed(apples)

#> ################## tWLRM ##################
#> 	The best estimate of total diversity is 1179 
#>  	 with std error 28 

wlrm_transformed(apples, plot = FALSE, print = FALSE, answers = TRUE)

#> $name
#> [1] "tWLRM"
#> 
#> $para
#>              Estimate Std. Error
#> (Intercept) 0.4388526 0.12855085
#> xs          0.1382131 0.01625013
#> 
#> $est
#> [1] 1178.603
#> 
#> $seest
#> [1] 28.17513
#> 
#> $full
#> 
#> Call:
#> lm(formula = log(ys) ~ xs, weights = weights_trans)
#> 
#> Coefficients:
#> (Intercept)           xs  
#>      0.4389       0.1382  
#> 
#> 
#> $ci
#> [1] 1013.925 3290.816
#>