function for species richness estimation

This function implements the untransformed version of the species richness estimation procedure outlined in Rocchetti, Bunge and Bohning (2011).

wlrm_untransformed(data, print = TRUE, plot = FALSE, answers = FALSE,
  cutoff = NA)

Arguments

data
The sample frequency count table for the population of interest. See breakaway for details.

print
Logical: whether the results should be printed to screen.

plot
Logical: whether the data and model fit should be plotted.

answers
Logical: whether the function should return an argument.

cutoff
The maximum frequency count to use for prediction. Defaults to maximal.

Value

name

The ``name'' of the model.

para

Estimated model parameters and standard errors.

est

The estimate of total (observed plus unobserved) diversity.

seest

The standard error in the diversity estimate.

full

The chosen linear model for frequency ratios.

ci

An asymmetric 95% confidence interval for diversity.

Note

This estimator is based on the negative binomial model and for that reason generally produces poor fits to microbial data. The result is usually artificially low standard errors. Caution is advised.

References

Rocchetti, I., Bunge, J. and Bohning, D. (2011). Population size estimation based upon ratios of recapture probabilities. Annals of Applied Statistics, 5.

See also

breakaway; apples; wlrm_transformed

Examples

wlrm_untransformed(apples)

#> ################## utWLRM ##################
#> 	The best estimate of total diversity is 1330 
#>  	 with std error 77